refactor: move Day to features
