+++ src/features/day/index.js
@@ -1,0 +1,1 @@
+export {Day} from './ui/Day';
